+++ src/backend/access/rmgrdesc/standbydesc.c
@@ -21,15 +21,15 @@ standby_desc_running_xacts(StringInfo buf, xl_running_xacts *xlrec)
 {
 	int			i;
 
-	appendStringInfo(buf, "nextXid %llu latestCompletedXid %llu oldestRunningXid %llu",
-					 (unsigned long long) xlrec->nextXid,
-					 (unsigned long long) xlrec->latestCompletedXid,
-					 (unsigned long long) xlrec->oldestRunningXid);
+	appendStringInfo(buf, "nextXid %" PRIu64 " latestCompletedXid %" PRIu64 " oldestRunningXid %" PRIu64,
+					 xlrec->nextXid,
+					 xlrec->latestCompletedXid,
+					 xlrec->oldestRunningXid);
 	if (xlrec->xcnt > 0)
 	{
 		appendStringInfo(buf, "; %d xacts:", xlrec->xcnt);
 		for (i = 0; i < xlrec->xcnt; i++)
-			appendStringInfo(buf, " %llu", (unsigned long long) xlrec->xids[i]);
+			appendStringInfo(buf, " %" PRIu64, xlrec->xids[i]);
 	}
 
 	if (xlrec->subxid_overflow)
@@ -39,8 +39,7 @@ standby_desc_running_xacts(StringInfo buf, xl_running_xacts *xlrec)
 	{
 		appendStringInfo(buf, "; %d subxacts:", xlrec->subxcnt);
 		for (i = 0; i < xlrec->subxcnt; i++)
-			appendStringInfo(buf, " %llu",
-							 (unsigned long long) xlrec->xids[xlrec->xcnt + i]);
+			appendStringInfo(buf, " %" PRIu64, xlrec->xids[xlrec->xcnt + i]);
 	}
 }
 
@@ -56,8 +55,8 @@ standby_desc(StringInfo buf, XLogReaderState *record)
 		int			i;
 
 		for (i = 0; i < xlrec->nlocks; i++)
-			appendStringInfo(buf, "xid %llu db %u rel %u ",
-							 (unsigned long long) xlrec->locks[i].xid,
+			appendStringInfo(buf, "xid %" PRIu64 " db %u rel %u ",
+							 xlrec->locks[i].xid,
 							 xlrec->locks[i].dbOid,
 							 xlrec->locks[i].relOid);
 	}
